refactoring directory structure, adding Branch_Name type
