+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import getBranches from "@/utils/api/getBranches";
+import getBranches from "@/utils/api/branches";
 import Landing from "@/components/Landing";
 
 export default async function Home() {
