--- src/interfaces/branch_name.ts
@@ -1,1 +1,0 @@
-export type Branch_Name = "navy" | "army" | "air-force" | "marine-corps";
