+++ src/types/custom.d.ts
@@ -1,0 +1,1 @@
+type Branch_Name = "navy" | "army" | "air-force" | "marine-corps";
