+++ src/utils/api/branches.ts
@@ -1,0 +1,5 @@
+export default async function getBranches() {
+  const res = await fetch(`${process.env.BASEURL}/api/branches`);
+
+  return res.json();
+}
